Upgrade ZIO Version (#435)
* upgrade zio version

* m4
