+++ zio-interop-cats-tests/shared/src/test/scala/zio/stream/interop/GenStreamInteropCats.scala
